Port to typescript-only static site generator
This was done mainly to remove the dependency on soupault. Soupault is a
very interesting tool, but has some serious problems:

- All scripting is done with embedded lua, and a very old version at
  that. This makes it impossible to intercept the scripts at runtime
  with any tool like a debugger.
- Soupault does not provide useful errors for any
  failure in the script (e.g. no line number and no description of the
  error).
- The "generation pipeline" of soupault makes it difficult to compute
  the rss feed and blog links once after building the index. These were
  being computed once per page in the website, which is of course very
  slow.

The poor error reporting and inability to debug lua with tools make
reasoning about what's happening in the scripting very difficult,
limiting the available approaches to spamming prints/logs and "the
squint test".